Use macro to select configuration To active the detctor we need to: echo 1 > /proc/hyplet_stats to deactivate it, toggle it: echo 1 > /proc/hyplet_stats
You can see that the detector is if you
cat /proc/hyplet_stats

However, when trying to set an ip on a deactivated interface
the system will hog. This is because it constantly peforms an
MMIO operating.

root@rpi3:~# [  222.684930] smsc95xx 1-1.1:1.0 eth0: Failed to read reg index 0x00000114: -110
[  222.692594] smsc95xx 1-1.1:1.0 eth0: Error reading MII_ACCESS
[  222.698697] smsc95xx 1-1.1:1.0 eth0: MII is busy in smsc95xx_mdio_read
[  222.705637] smsc95xx 1-1.1:1.0 eth0: Failed to read MII_BMSR
[  240.020805] INFO: rcu_sched detected stalls on CPUs/tasks:
[  240.026637]  0-...: (1 GPs behind) idle=f3e/140000000000001/0 softirq=1937/1938 fqs=2626
[  240.035238]  (detected by 3, t=5255 jiffies, g=706, c=705, q=374)
[  240.041654] Task dump for CPU 0:
[  240.045050] mmcqd/0         R  running task        0    89      2 0x00000022
[  240.052468] Call trace:
[  240.055056] [<ffffff84a68864ac>] __switch_to+0x94/0xa8
[  240.060462] [<00000000a68da150>] 0xa68da150

To fix that, we need to report access, fail the first MMIO, and then
turn the detector off.
All of the above relates only to the smc driver. It is unclear at the moment
how would USB drivers react.